fix: access requests nextjs server component error
