+++ client/app/(authenticated)/idir/cas_admin/dashboard/operators/page.tsx
@@ -1,11 +1,3 @@
-import { Suspense } from "react";
-import AccessRequests from "@/app/components/routes/access-requests/AccessRequests";
-import Loading from "@/app/components/loading/SkeletonGrid";
+import OperatorsPage from "@/app/components/routes/operators/OperatorsPage";
 
-export default function Page() {
-  return (
-    <Suspense fallback={<Loading />}>
-      <AccessRequests />
-    </Suspense>
-  );
-}
+export default OperatorsPage;
